Update upload() and download() methods
upload() was storing files in Firebase Storage with the same absolute path that the file had on the device. It now stores the file in Firebase Storage's root directory.

download() would only download files with the name "test.txt" and I could not find where the files were downloaded to. It now downloads whatever file name was entered if it exists and returns the absolute path of the downloaded file on the device in data.